Solution logout
