+++ app/Logout/logout_data/src/main/java/com/tatho/logout_data/reposiroty/AuthFirebaseLogoutRepositoryImpl.kt
@@ -1,0 +1,15 @@
+package com.tatho.logout_data.reposiroty
+
+import android.content.SharedPreferences
+import com.google.firebase.auth.FirebaseAuth
+import com.tatho.logout_domain.repository.IAuthFirebaseLogoutRepository
+
+class AuthFirebaseLogoutRepositoryImpl(
+    private var auth: FirebaseAuth,
+    private val sharedPreferences: SharedPreferences
+) : IAuthFirebaseLogoutRepository {
+    override suspend fun logout() {
+        auth.signOut()
+        sharedPreferences.edit().putString("Uid", "").apply()
+    }
+}
