+++ app/Logout/logout_domain/src/main/java/com/tatho/logout_domain/repository/IAuthFirebaseLogoutRepository.kt
@@ -1,0 +1,5 @@
+package com.tatho.logout_domain.repository
+
+interface IAuthFirebaseLogoutRepository {
+    suspend fun logout()
+}
