+++ app/Logout/logout_presentation/build.gradle
@@ -41,11 +41,10 @@ android {
 
 dependencies {
 
+
     implementation 'androidx.core:core-ktx:1.7.0'
 
     implementation project(path: ':app:Common')
-    implementation project(path: ':app:RegisterSize:domain')
-    implementation project(path: ':app:SingUp:sing_domain')
 
     def composeBom = platform('androidx.compose:compose-bom:2023.05.01')
     implementation composeBom
