+++ app/Menu/menu_data/build.gradle
@@ -34,10 +34,10 @@ android {
 dependencies {
 
     implementation project(path: ':app:RegisterSize:domain')
+    implementation project(path: ':app:Menu:menu_domain')
 
     implementation 'androidx.core:core-ktx:1.10.1'
     implementation "com.google.dagger:hilt-android:2.44"
-    implementation project(path: ':app:Menu:menu_domain')
     kapt "com.google.dagger:hilt-compiler:2.44"
 
     //Room
