+++ app/Menu/menu_presentation/build.gradle
@@ -46,6 +46,8 @@ dependencies {
     implementation project(path: ':app:Common')
     implementation project(path: ':app:Menu:menu_domain')
     implementation project(path: ':app:Menu:menu_data')
+    implementation project(path: ':app:Logout:logout_domain')
+    implementation project(path: ':app:Logout:logout_presentation')
 
     def composeBom = platform('androidx.compose:compose-bom:2023.05.01')
     implementation composeBom
