+++ app/Menu/menu_presentation/src/main/java/com/tatho/presentation/components/ItemMenuTitle.kt
@@ -1,0 +1,18 @@
+package com.tatho.presentation.components
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun ItemMenuTitle(title: String) {
+    Text(
+        text = title,
+        fontSize = 16.sp,
+        fontWeight = FontWeight.Bold,
+        color = Color(0xFF2D3338),
+        letterSpacing = 0.25.sp,
+    )
+}
