+++ app/RegisterSize/domain/src/main/java/com/tatho/domain/di/BodyMeasurementDomainModule.kt
@@ -9,7 +9,6 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module
