+++ app/SingUp/sing_data/build.gradle
@@ -33,14 +33,13 @@ android {
 
 dependencies {
 
+    implementation project(path: ':app:SingUp:sing_domain')
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.9.0'
-    implementation project(path: ':app:SingUp:sing_domain')
     implementation 'androidx.core:core-ktx:1.10.1'
     implementation "com.google.dagger:hilt-android:2.44"
-    implementation 'com.google.firebase:firebase-database-ktx:20.1.0'
     kapt "com.google.dagger:hilt-compiler:2.44"
 
     //Room
@@ -50,7 +49,7 @@ dependencies {
     implementation 'com.google.firebase:firebase-firestore-ktx:24.4.1'
     implementation platform('com.google.firebase:firebase-bom:32.1.1')
     implementation 'com.google.firebase:firebase-analytics-ktx'
-
+    implementation 'com.google.firebase:firebase-database-ktx:20.1.0'
     // Import the BoM for the Firebase platform
     implementation(platform("com.google.firebase:firebase-bom:32.1.1"))
