--- app/SingUp/sing_data/src/main/java/com/tatho/sing_data/di/FirebaseAuth.kt
@@ -1,26 +1,0 @@
-package com.tatho.sing_data.di
-
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
-import com.tatho.sing_data.repository.AuthFirebaseRepoImpl
-import com.tatho.sing_domain.repository.IAuthFirebaseRepository
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-class FirebaseAuthModel {
-
-    @Provides
-    fun provideFirebaseAuth(): FirebaseAuth {
-        return Firebase.auth
-    }
-
-    @Provides
-    fun provideAuthFirebaseRepo(firebaseAuth: FirebaseAuth): IAuthFirebaseRepository {
-        return AuthFirebaseRepoImpl(firebaseAuth)
-    }
-}
