+++ app/SingUp/sing_data/src/main/java/com/tatho/sing_data/repository/AuthFirebaseSingUpRepoImpl.kt
@@ -1,17 +1,20 @@
 package com.tatho.sing_data.repository
 
 
+import android.content.SharedPreferences
 import android.util.Log
 import com.google.firebase.auth.FirebaseAuth
-import com.tatho.sing_domain.repository.IAuthFirebaseRepository
+import com.tatho.sing_domain.repository.IAuthFirebaseSingUpRepository
 
-class AuthFirebaseRepoImpl(
-    private var auth: FirebaseAuth
-) : IAuthFirebaseRepository {
+class AuthFirebaseSingUpRepoImpl(
+    private var auth: FirebaseAuth,
+    private val sharedPreferences: SharedPreferences,
+) : IAuthFirebaseSingUpRepository {
     override suspend fun register(email: String, password: String, callback: (String) -> Unit) {
         auth.createUserWithEmailAndPassword(email, password)
             .addOnSuccessListener {
                 it.user?.uid?.let { uid ->
+                    sharedPreferences.edit().putString("Uid", uid).apply()
                     callback(uid)
                 }
                 Log.e("AUT", "Se creo el usuario")
