+++ app/SingUp/sing_domain/build.gradle
@@ -35,34 +35,6 @@ dependencies {
 
     implementation 'androidx.core:core-ktx:1.10.1'
     implementation 'androidx.appcompat:appcompat:1.6.1'
-//    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-
-//    testImplementation 'junit:junit:4.13.2'
-//    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-//    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-
-//    implementation 'com.airbnb.android:lottie:3.4.2'
-//
-//    implementation 'com.google.dagger:hilt-android:2.46.1'
-//    kapt 'com.google.dagger:hilt-android-compiler:2.44'
-//    kapt 'androidx.hilt:hilt-compiler:1.0.0'
-//
-//    implementation 'androidx.room:room-ktx:2.4.3'
-//    kapt 'androidx.room:room-compiler:2.4.3'
-//
-//    def compose_version = "1.4.0"
-//    implementation "androidx.compose.ui:ui:$compose_version"
-//    implementation "androidx.compose.ui:ui-tooling-preview:$compose_version"
-//    debugImplementation "androidx.compose.ui:ui-tooling:$compose_version"
-//    debugImplementation "androidx.compose.ui:ui-test-manifest:$compose_version"
-//    implementation "androidx.compose.runtime:runtime-livedata:$compose_version"
-//
-//    def nav_version = "2.4.0-alpha05"
-//    implementation "androidx.navigation:navigation-compose:$nav_version"
-//
-//    implementation 'com.google.android.material:material:1.9.0'
-
-    //Room
     implementation 'androidx.room:room-ktx:2.5.2'
     implementation "com.google.dagger:hilt-android:2.44"
     implementation project(path: ':app:Common')
