+++ app/SingUp/sing_domain/src/main/java/com/tatho/sing_domain/repository/IAuthFirebaseSingUpRepository.kt
@@ -1,5 +1,5 @@
 package com.tatho.sing_domain.repository
 
-interface IAuthFirebaseRepository {
+interface IAuthFirebaseSingUpRepository {
     suspend fun register(email: String, password: String, callback: (String) -> Unit)
 }
