+++ settings.gradle
@@ -27,3 +27,7 @@ include ':app:SingUp'
 include ':app:SingUp:sing_data'
 include ':app:SingUp:sing_domain'
 include ':app:SingUp:sing_presentation'
+include ':app:Logout'
+include ':app:Logout:logout_data'
+include ':app:Logout:logout_domain'
+include ':app:Logout:logout_presentation'
